+++ CHANGELOG.md
@@ -7,12 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - MinimumScoreFilter can have a null value in a ValueGroupResult
+- Warning when building an Extract request from a stream resource without an associated URI/filename
 
 
 ## [6.3.1]
 ### Added
 - Loadbalancer plugin can failover on an optional list of HTTP status codes
-- Solarium\QueryType\Extract\Query::setFile() now supports file pointer resources
+- Solarium\QueryType\Extract\Query::setFile() now supports stream resources
 - Solarium\QueryType\Extract\Result::getFile() and getFileMetadata() to access the retrieved data for `extractOnly=true`
 
 ### Fixed