+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Solarium library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [6.3.2]
 ### Added
 - Solarium\Component\ReRankQuery::setOperator()
 - Solarium\Component\ReRankQuery::setScale() and setMainScale()